requirements.txt
Faker==35.2.0
kafka-python==2.0.2
pyspark==3.5.4